Initialize UI and Add Constants and Various Functionality.
The Server UI is planned to have a few important nodes. The server
should be able to display user commands, the connected clients, the
communication log between the client and server, and lastly have the
ability to send commands to the client. Additional functionality may be
expected in the future.