Initialize Client Main Loop and Refactor Code
The main loop will listen inside the FSClient is planned to listen for
data that is sent between the server and client. Right now it only
checks if the client has disconnected. When the client disconnects
then close the IO streams and the socket connection. Each client is
also disconnected when the server closes. Added some log messages to
the console to let the user know.